pyproject.toml
[tool.poetry]
name = "repl_python3_Euclidean"
version = "0.1.0"
description = ""
authors = ["PrithwishNag <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.1"
tabulate = "^0.8.9"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
